server check

+++ TankUIClient/Assets/Script/GameManager.cs
@@ -194,12 +194,12 @@ public class GameManager : MonoBehaviour
         if(team1Score == 3 || team2Count == 0)
         {
             isEnd = 1;
-            PacketSender.WinGameSender(Client.instance.roomId, (byte)0);
+            PacketSender.WinGameSender(Client.instance.roomId, 0);
         }
         else if(team2Score == 3 || team1Count == 0)
         {
             isEnd = 1;
-            PacketSender.WinGameSender(Client.instance.roomId, (byte)1);
+            PacketSender.WinGameSender(Client.instance.roomId, 1);
         }
 
     }
